Permissions added

+++ routers/database.py
@@ -15,8 +15,10 @@ from config.database import SessionLocal
 database_router = APIRouter()
 
 # Desc: Function to create a new database record calling the service class.
-@database_router.post('/database/create', tags=['Database'], response_model=DatabaseSchema, summary='Create a new database record - Crear un nuevo registro en la base de datos.')
-def create_new_database_record(database: DatabaseSchema = Depends()):
+@database_router.post('/database/create', tags=['Database'], response_model=DatabaseSchema, summary='Create a new database record - Crear un nuevo registro en la base de datos.', dependencies=[Depends(oauth2_scheme)])
+def create_new_database_record(database: DatabaseSchema = Depends(),role: str = Depends(get_current_active_user_role)):
+    if role not in ['Admin', 'User']:
+        return JSONResponse(status_code=403, content={'message': 'Error: You do not have permission to perform this action.'})
     try:
         db = SessionLocal()
         check_tin = DatabaseServices(db).get_by_tin(database.tin)
@@ -33,8 +35,10 @@ def create_new_database_record(database: DatabaseSchema = Depends()):
         raise HTTPException(status_code=500, detail='Internal server error.')
 
 # Desc: Function to get the database records calling the service class.
-@database_router.get('/database/get', tags=['Database'], response_model=List[DatabaseSchema], summary='Get all the database records - Obtener todos los registros de la base de datos.')
-def get_all_database_records():
+@database_router.get('/database/get', tags=['Database'], response_model=List[DatabaseSchema], summary='Get all the database records - Obtener todos los registros de la base de datos.', dependencies=[Depends(oauth2_scheme)])
+def get_all_database_records(role: str = Depends(get_current_active_user_role)):
+    if role not in ['Admin', 'User']:
+        return JSONResponse(status_code=403, content={'message': 'Error: You do not have permission to perform this action.'})
     try:    
         db = SessionLocal()
         records = DatabaseServices(db).get_all()
@@ -47,8 +51,10 @@ def get_all_database_records():
         raise HTTPException(status_code=500, detail='Internal server error.')
     
 # Desc: Function to update a database record by TIN calling the service class.
-@database_router.put('/database/update/{tin_to_search}', tags=['Database'], response_model=UpdateSchema, summary='Update a database record by TIN - Actualizar un registro de la base de datos por NIT.')
-def update_database_record_by_tin(tin_to_search: int = Path(..., title='TIN', description='TIN of the record to be updated. - NIT del registro a actualizar.', gt=0), database: UpdateSchema = Depends()):
+@database_router.put('/database/update/{tin_to_search}', tags=['Database'], response_model=UpdateSchema, summary='Update a database record by TIN - Actualizar un registro de la base de datos por NIT.', dependencies=[Depends(oauth2_scheme)])
+def update_database_record_by_tin(tin_to_search: int = Path(..., title='TIN', description='TIN of the record to be updated. - NIT del registro a actualizar.', gt=0), database: UpdateSchema = Depends(), role: str = Depends(get_current_active_user_role)):
+    if role not in ['Admin']:
+        return JSONResponse(status_code=403, content={'message': 'Error: You do not have permission to perform this action.'})
     try:
         db = SessionLocal()
         record_to_update = DatabaseServices(db).update_by_tin(tin_to_search, database)
@@ -63,8 +69,10 @@ def update_database_record_by_tin(tin_to_search: int = Path(..., title='TIN', de
         raise HTTPException(status_code=500, detail='Internal server error.')
     
 # Desc: Function to eliminate a database record by TIN calling the service class.
-@database_router.delete('/database/delete/{tin}', tags=['Database'], summary='Delete a database record by TIN - Eliminar un registro de la base de datos por NIT.')
-def delete_database_record_by_tin(tin: int = Path(..., title='TIN', description='TIN of the record to be deleted. - NIT del registro a eliminar.', gt=0)):
+@database_router.delete('/database/delete/{tin}', tags=['Database'], summary='Delete a database record by TIN - Eliminar un registro de la base de datos por NIT.', dependencies=[Depends(oauth2_scheme)])
+def delete_database_record_by_tin(tin: int = Path(..., title='TIN', description='TIN of the record to be deleted. - NIT del registro a eliminar.', gt=0), role: str = Depends(get_current_active_user_role)):
+    if role not in ['Admin']:
+        return JSONResponse(status_code=403, content={'message': 'Error: You do not have permission to perform this action.'})
     try:
         db = SessionLocal()
         record_to_delete = DatabaseServices(db).delete_by_tin(tin)
